Create README.md
Im phantom5183 and Im interested in helping build, operate & manage a prosperous environment for everyone. Im in the learning stage and am open to many ideas and hope to be able to impart as much as I learn. I'm looking to collaborate on any projects that encompasses that described in my interests. U can reach me at #phantom5183.